UNSTABLE: merged version corrected
However, widows and orphans problem is back.
May need to check previous sources to get
all changes in.

+++ src/main/java/org/mapfish/print/map/renderers/TestTileRenderer.java
@@ -16,10 +16,9 @@ public class TestTileRenderer extends TileRenderer {
 	public void render(Transformer transformer, List<URI> urls,
 			ParallelMapTileLoader parallelMapTileLoader,
 			RenderingContext context, float opacity, int nbTilesHorizontal,
-			float offsetX, float offsetY, long bitmapTileW, long bitmapTileH)
+			double offsetX, double offsetY, long bitmapTileW, long bitmapTileH)
 			throws IOException {
 		lastURIs = urls;
-
 	}
 
 }
